services/auth/requirements.txt
flask
flask-cors
bcrypt
pyjwt
psycopg2-binary
sqlalchemy
